+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.1] - 2021-04-06
+### Added
+- Rest framework ``serializers.ChoiceField`` and ``serializers.MultipleChoiceField`` will now be represented as ``Enum`` types with enum values correctly reflected in the schema.
+- Documentation for using Tags.
+
+### Fixed
+- Fix bug where schema examples are not generated correctly.
+- Fix bug where the request URL for the objects are generated with an incorrect prefix.
 
 ## [1.1.0] - 2021-01-04
 ### Added